fix(office): keep media result contract dependency-light
