+++ packages/coding-agent/src/media/publish.ts
@@ -1,16 +1,11 @@
 import type { AgentToolResult } from "@f5-sales-demo/pi-agent-core";
 import type { ImageContent, TextContent } from "@f5-sales-demo/pi-ai";
 import type { IngestedMedia } from "./ingest";
-import type { MediaDescriptorV1, MediaMessage } from "./types";
+import { MEDIA_TOOL_RESULT_V1, type MediaToolResultV1, validateMediaToolResultV1 } from "./tool-result";
+import type { MediaMessage } from "./types";
 import { validateMediaDescriptorV1 } from "./types";
 
-export const MEDIA_TOOL_RESULT_V1 = "xcsh.media/v1" as const;
-
-export interface MediaToolResultV1 {
-	mediaResult: typeof MEDIA_TOOL_RESULT_V1;
-	descriptor: MediaDescriptorV1;
-	displayMethod: "inline" | "poster" | "text";
-}
+export { MEDIA_TOOL_RESULT_V1, type MediaToolResultV1, validateMediaToolResultV1 } from "./tool-result";
 
 export interface MediaPublisherSession {
 	appendMediaMessage?: (message: MediaMessage) => void;
@@ -19,18 +14,6 @@ export interface MediaPublisherSession {
 export interface PublishMediaOptions {
 	text?: string[];
 	timestamp?: number;
-}
-
-export function validateMediaToolResultV1(value: unknown): MediaToolResultV1 {
-	if (!value || typeof value !== "object") throw new Error("media tool result must be an object");
-	const result = value as Partial<MediaToolResultV1>;
-	if (result.mediaResult !== MEDIA_TOOL_RESULT_V1) throw new Error("unsupported media tool result version");
-	if (!result.descriptor) throw new Error("media tool result requires a descriptor");
-	const descriptor = validateMediaDescriptorV1(result.descriptor);
-	if (!(["inline", "poster", "text"] as const).includes(result.displayMethod as never)) {
-		throw new Error("invalid media display method");
-	}
-	return { mediaResult: MEDIA_TOOL_RESULT_V1, descriptor, displayMethod: result.displayMethod! };
 }
 
 /**
